backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
langgraph==0.0.20
groq==0.4.2
langchain==0.1.0
langchain-groq==0.0.1
pypdf2==3.0.1
python-docx==1.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
